Fixed multiple errors

+++ trunk/MM/functions.php
@@ -1,23 +1,72 @@
 <?php
 /**
+ * MidiSL page include function
+ * @param string $name name of the page to be included
+ * @param array $namesAndPages Array mapping pagenames to actual pages
+ */
+function midiInclude($name, $namesAndPages) {
+	include($namesAndPages[$name]);
+}
+/**
  * Function to delete a single record from the database
  * @param int $id the transaction id of the record that needs to be deleted
+ * @param string $page the calling page
  * @return bool whether the deletion process was succesful
  */
-function deleteRecord($id) {
-   $transDeleted = mysql_query("DELETE FROM pawning WHERE ref_no='$id'");
-   $refDelete = mysql_query("DELETE FROM customer_ref WHERE ref_no='$id'");
-   if ( $transDeleted && $refDelete ) {
-       return true;
-   }
-   
+function deleteRecord($id, $page) {
+	if ( $page == 'pawning' ) {
+		$transDeleted = mysql_query("DELETE FROM pawning WHERE ref_no='$id'");
+		$refDelete = mysql_query("DELETE FROM customer_ref WHERE ref_no='$id'");
+		if ( $transDeleted && $refDelete ) {
+			return true;
+		}
+	}
+	if ( $page == 'redeem' ) {
+		$redDeleted = mysql_query("DELETE FROM redeems WHERE ref_no='$id'");
+		if ( $redDeleted ) {
+			return true;
+		}
+	}
+	if ( $page == 'expences' ) {
+		$exDeleted = mysql_query("DELETE FROM expenses WHERE ex_id='$id'");
+		if ( $exDeleted ) {
+			return true;
+		}
+	}
+	if ( $page == 'sinna' ) {
+		$sinnaDeleted = mysql_query("DELETE FROM sinna WHERE ref_no='$id'");
+		if ( $sinnaDeleted ) {
+			return true;
+		}
+	}
+	if ( $page == 'deposit' ) {
+		$depositDeleted = mysql_query("DELETE FROM deposit WHERE trans_id='$id'");
+		if ( $depositDeleted ) {
+			return true;
+		}
+	}
+	if ( $page == 'withdraw' ) {
+		$wdDeleted = mysql_query("DELETE FROM withdrawals WHERE trans_id='$id'");
+		if ( $wdDeleted ) {
+			return true;
+		}
+	}
+	if ( $page == 'view' ) {
+		$transDeleted = mysql_query("DELETE FROM pawning WHERE ref_no='$id'");
+		$refDelete = mysql_query("DELETE FROM customer_ref WHERE ref_no='$id'");
+		if ( $transDeleted && $refDelete ) {
+			return true;
+		}
+	}
+
 }
 /**
  * Display function to get data from tables and display
  * @param string $page page which is calling the function in a way that can be passed to $_GET
  */
 function display($page) {
-    echo '<frameset><legend>Details</legend>
+
+	echo '<frameset><legend>Details</legend>
         <table width="80%" border="1">
 		<tr>
         <th>No</th>
@@ -28,30 +77,34 @@ function display($page) {
 		<th>Type</th>
 		<th>Branch</th>
 		<th>Delete</th>
-	</tr>';
-    $viewQ = "SELECT * FROM pawning LIMIT 20";
-    $view = mysql_query($viewQ);
-    $num = mysql_num_rows($view);
-    $no=1;
-    for ( $i = 0; $i < $num; $i++ ) {
-	$ref = mysql_result($view,$i,'ref_no');
-	echo '<tr><td>'.$no++.'</td>'.
+		</tr>';
+	$viewQ = "SELECT * FROM pawning LIMIT 20";
+	$view = mysql_query($viewQ);
+	$num = mysql_num_rows($view);
+	$no=1;
+	for ( $i = 0; $i < $num; $i++ ) {
+		$ref = mysql_result($view,$i,'ref_no');
+		echo '<tr><td>'.$no++.'</td>'.
                  '<td>'.$ref.'</td>'.
-		 '<td>'.mysql_result($view,$i,'date').'</td>'.
-		 '<td>'.mysql_result($view,$i,'weight').'</td>'.
-		 '<td>'.mysql_result($view,$i,'amount').'</td>'.
-		 '<td>'.mysql_result($view,$i,'type').'</td>'.
-		 '<td>'.mysql_result($view,$i,'branch').'</td>'.
-		 '<td style="text-align:center"><a href="home.php?page='.$page.'&func=delete&ref='.$ref.'"><img src="images/b_drop.png" /></a></td></tr>';
-    }
-    echo '</table></frameset>';
+		 		 '<td>'.mysql_result($view,$i,'date').'</td>'.
+		 		 '<td>'.mysql_result($view,$i,'weight').'</td>'.
+		 		 '<td>'.mysql_result($view,$i,'amount').'</td>'.
+		  		 '<td>'.mysql_result($view,$i,'type').'</td>'.
+		 		 '<td>'.mysql_result($view,$i,'branch').'</td>'.
+		 		 '<td style="text-align:center"><a href="home.php?page='.$page.'&func=delete&ref='.$ref.'"><img src="images/b_drop.png" /></a></td></tr>';
+		
+	}
+	echo '</table></frameset>';
 }
+
+
 /**
  * Display function to get today's transactions and display them
  * @param string $page the calling page
  */
 function displayToday($page) {
-    echo '<frameset><legend>Details</legend>
+	if( $page == pawning ){
+		echo '<frameset><legend>Details</legend>
         <table width="80%" border="1">
 	<tr>
         <th>No</th>
@@ -63,15 +116,15 @@ function displayToday($page) {
 		<th>Branch</th>
 		<th>Delete</th>
 	</tr>';
-    $branch = $_SESSION['branch'];
-    $today = date('Y-m-d');
-    $viewQ = "SELECT * FROM pawning WHERE date='$today' AND branch='$branch'";
-    $view = mysql_query($viewQ);
-    $num = mysql_num_rows($view);
-    $no=1;
-    for ( $i = 0; $i < $num; $i++ ) {
-	$ref = mysql_result($view,$i,'ref_no');
-	echo '<tr><td>'.$no++.'</td>'.
+		$branch = $_SESSION['branch'];
+		$today = date('Y-m-d');
+		$viewQ = "SELECT * FROM pawning WHERE date='$today' AND branch='$branch'";
+		$view = mysql_query($viewQ);
+		$num = mysql_num_rows($view);
+		$no=1;
+		for ( $i = 0; $i < $num; $i++ ) {
+			$ref = mysql_result($view,$i,'ref_no');
+			echo '<tr><td>'.$no++.'</td>'.
                  '<td>'.$ref.'</td>'.
 		 '<td>'.mysql_result($view,$i,'date').'</td>'.
 		 '<td>'.mysql_result($view,$i,'weight').'</td>'.
@@ -79,96 +132,144 @@ function displayToday($page) {
 		 '<td>'.mysql_result($view,$i,'type').'</td>'.
 		 '<td>'.mysql_result($view,$i,'branch').'</td>'.
 		 '<td style="text-align:center"><a href="home.php?page='.$page.'&func=delete&ref='.$ref.'"><img src="images/b_drop.png" /></a></td></tr>';
-    }
-    echo '</table></frameset>';
+		}
+		echo '</table></frameset>';
+	}
+
+	if( $page == withdraw ){
+		echo '<frameset><legend>Details</legend>
+        <table width="80%" border="1">
+	<tr>
+        <th>No</th>
+		<th>Source</th>
+		<th>Amount</th>
+		<th>Description</th>
+		<th>Date</th>
+	</tr>';
+		$branch = $_SESSION['branch'];
+		$today = date('Y-m-d');
+		$viewQ = "SELECT * FROM withdrawals WHERE date='$today'";
+		$view = mysql_query($viewQ);
+		$num = mysql_num_rows($view);
+		$no=1;
+		for ( $i = 0; $i < $num; $i++ ) {
+			//$ref = mysql_result($view,$i,'ref_no');
+			echo '<tr><td>'.$no++.'</td>'.
+                 '<td>'.$ref.'</td>'.
+		 '<td>'.mysql_result($view,$i,'trans_id').'</td>'.
+		 '<td>'.mysql_result($view,$i,'source').'</td>'.
+		 '<td>'.mysql_result($view,$i,'amount').'</td>'.
+		 '<td>'.mysql_result($view,$i,'description').'</td>'.
+		 '<td>'.mysql_result($view,$i,'date').'</td>'.
+		 '<td style="text-align:center"><a href="home.php?page='.$page.'&func=delete&ref='.$ref.'"><img src="images/b_drop.png" /></a></td></tr>';
+		}
+		echo '</table></frameset>';
+	}
+
+	if( $page == expenses ){
+		echo '<frameset><legend>Details</legend>
+        <table width="80%" border="1">
+	<tr>
+        <th>ID</th>
+		<th>Branch</th>
+		<th>Date</th>
+		<th>Discription</th>
+		<th>Amount</th>
+		<th>Delete</th>
+	</tr>';
+		$branch = $_SESSION['branch'];
+		$today = date('Y-m-d');
+		$viewQ = "SELECT * FROM expenses WHERE date='$today'";
+		$view = mysql_query($viewQ);
+		$num = mysql_num_rows($view);
+		$no=1;
+		for ( $i = 0; $i < $num; $i++ ) {
+			//$ref = mysql_result($view,$i,'ref_no');
+			echo '<tr>
+		  <td>'.mysql_result($view,$i,'ex_id').'</td>'.
+		 '<td>'.mysql_result($view,$i,'branch').'</td>'.
+		 '<td>'.mysql_result($view,$i,'date').'</td>'.
+		 '<td>'.mysql_result($view,$i,'discription').'</td>'.
+		 '<td style="text-align:center"><a href="home.php?page='.$page.'&func=delete&ref='.$ref.'"><img src="images/b_drop.png" /></a></td></tr>';
+		}
+		echo '</table></frameset>';
+	}
 }
+
 /**
  * Function to return a full data set from pawning table for stats
  * @param string $page the calling pagename as should be passed to $_GET
  * @return array $data array containing all data
  */
 function showStats($page) {
-    $date = date('Y-m-d');
-    $data = array();
-    $data['total'] = array();
-    $data['anu_market'] = array();
-    $data['anu_bus'] = array();
-    $data['kekirawa'] = array();
-    $data['mm'] = array();
-    $data['wariyapola'] = array();
-    $data['kurunegala'] = array();
-    $data['medawachchiya'] = array();
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date'"));
-    $data['total']['trans'] = $temp['COUNT(*)'];
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='kekirawa'"));
-    $data['kekirawa']['trans'] = $temp['COUNT(*)'];
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='anu_market'"));
-    $data['anu_market']['trans'] = $temp['COUNT(*)']; 
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='anu_bus'"));
-    $data['anu_bus']['trans'] = $temp['COUNT(*)'];
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='mm'"));
-    $data['mm']['trans'] = $temp['COUNT(*)'];
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='kurunegala'"));
-    $data['kurunegala']['trans'] = $temp['COUNT(*)'];
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='wariyapola'"));
-    $data['wariyapola']['trans'] = $temp['COUNT(*)'];
-    $temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='medawachchiya'"));
-    $data['medawachchiya']['trans'] = $temp['COUNT(*)'];
-    
-    $gold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date'"));
-    
-    $data['total']['pawning'] = $gold['pawns'];
-    $data['total']['gold'] = $gold['gold'];
-    
-    $kekirawaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='kekirawa'"));
-    
-    $data['kekirawa']['pawning'] = $kekirawaGold['pawns'];
-    $data['kekirawa']['gold'] = $kekirawaGold['gold'];
-    
-    $anuMarketGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='anu_market'"));
-    
-    $data['anu_market']['pawning'] = $anuMarketGold['pawns'];
-    $data['anu_market']['gold'] = $anuMarketGold['gold'];
-    
-    $anuBusGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='anu_bus'"));
-    
-    $data['anu_bus']['pawning'] = $anuBusGold['pawns'];
-    $data['anu_bus']['gold'] = $anuBusGold['gold'];
-    
-    $mmGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='mm'"));
-    
-    $data['mm']['pawning'] = $mmGold['pawns'];
-    $data['mm']['gold'] = $mmGold['gold'];
-    
-    $kurunegalaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='kurunegala'"));
-    
-    $data['kurunegala']['pawning'] = $mmGold['pawns'];
-    $data['kurunegala']['gold'] = $mmGold['gold'];
-    
-    $wariyapolaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='wariyapola'"));
-    
-    $data['wariyapola']['pawning'] = $wariyapolaGold['pawns'];
-    $data['wariyapola']['gold'] = $wariyapolaGold['gold'];
-    
-    $medawachchiyaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='medawachchiya'"));
-    
-    $data['medawachchiya']['pawning'] = $medawachchiyaGold['pawns'];
-    $data['medawachchiya']['gold'] = $medawachchiyaGold['gold'];
-    
-    /*$display = '<h3>Today\'s transactions</h3><br/><table border="0" width="80%">'
-                .'<tr><td>Total transactions</td><td>'.$pawnNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Anuradhapura Market transactions</td><td>'.$anuMarketNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Anuradhapura Bus Stand transactions</td><td>'.$anuBusNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Kekirawa transactions</td><td>'.$kekirawaNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>MM transactions</td><td>'.$mmNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Kurunegala transactions</td><td>'.$kurunegalaNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Wariyapola transactions</td><td>'.$wariyapolaNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Medawachchiya transactions</td><td>'.$medawachchiyaNumber['COUNT(*)'].'</td></tr>'
-                .'<tr><td>Total gold weight</td><td>'.$totalGold.'</td></tr>'
-                .'</table>';
-    echo $display;*/
-    //echo $data['total']['pawning'];
-    return $data;
+	$date = date('Y-m-d');
+	$data = array();
+	$data['total'] = array();
+	$data['anu_market'] = array();
+	$data['anu_bus'] = array();
+	$data['kekirawa'] = array();
+	$data['mm'] = array();
+	$data['wariyapola'] = array();
+	$data['kurunegala'] = array();
+	$data['medawachchiya'] = array();
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date'"));
+	$data['total']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='kekirawa'"));
+	$data['kekirawa']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='anu_market'"));
+	$data['anu_market']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='anu_bus'"));
+	$data['anu_bus']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='mm'"));
+	$data['mm']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='kurunegala'"));
+	$data['kurunegala']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='wariyapola'"));
+	$data['wariyapola']['trans'] = $temp['COUNT(*)'];
+	$temp = mysql_fetch_assoc(mysql_query("SELECT COUNT(*) FROM pawning WHERE date='$date' AND branch='medawachchiya'"));
+	$data['medawachchiya']['trans'] = $temp['COUNT(*)'];
+
+	$gold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date'"));
+
+	$data['total']['pawning'] = $gold['pawns'];
+	$data['total']['gold'] = $gold['gold'];
+
+	$kekirawaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='kekirawa'"));
+
+	$data['kekirawa']['pawning'] = $kekirawaGold['pawns'];
+	$data['kekirawa']['gold'] = $kekirawaGold['gold'];
+
+	$anuMarketGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='anu_market'"));
+
+	$data['anu_market']['pawning'] = $anuMarketGold['pawns'];
+	$data['anu_market']['gold'] = $anuMarketGold['gold'];
+
+	$anuBusGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='anu_bus'"));
+
+	$data['anu_bus']['pawning'] = $anuBusGold['pawns'];
+	$data['anu_bus']['gold'] = $anuBusGold['gold'];
+
+	$mmGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='mm'"));
+
+	$data['mm']['pawning'] = $mmGold['pawns'];
+	$data['mm']['gold'] = $mmGold['gold'];
+
+	$kurunegalaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='kurunegala'"));
+
+	$data['kurunegala']['pawning'] = $mmGold['pawns'];
+	$data['kurunegala']['gold'] = $mmGold['gold'];
+
+	$wariyapolaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='wariyapola'"));
+
+	$data['wariyapola']['pawning'] = $wariyapolaGold['pawns'];
+	$data['wariyapola']['gold'] = $wariyapolaGold['gold'];
+
+	$medawachchiyaGold = mysql_fetch_assoc(mysql_query("SELECT SUM(amount) AS pawns,SUM(weight) AS gold FROM pawning WHERE date='$date' AND branch='medawachchiya'"));
+
+	$data['medawachchiya']['pawning'] = $medawachchiyaGold['pawns'];
+	$data['medawachchiya']['gold'] = $medawachchiyaGold['gold'];
+
+	return $data;
 }
 /**
  * Function called by getEmp.php to get employee details (related to showEmp() ajax function)
@@ -197,7 +298,7 @@ function dispEmp($name) {
  * @param string $name name of the customer
  */
 function dispRefs($name) {
-	$cusId = mysql_fetch_assoc(mysql_query("SELECT cus_id FROM customer_details WHERE name='$name'"));
+	$cusId = mysql_fetch_assoc(mysql_query("SELECT cus_id FROM customer_details WHERE name='$name' OR name LIKE '$name%' OR name LIKE '%$name' OR name LIKE '%$name%'"));
 	$cusId = $cusId['cus_id'];
 	$refs = mysql_query("SELECT ref_no FROM customer_ref WHERE cus_id='$cusId'");
 	$num = mysql_num_rows($refs);
@@ -211,12 +312,19 @@ function dispRefs($name) {
 }
 
 function dispSingle($id) {
+	session_start();
+	$_SESSION['ref_no'] = $id;
 	$details = mysql_fetch_assoc(mysql_query("SELECT * FROM pawning WHERE ref_no='$id'"));
-	echo '<table width="70%" border="0">';
-	echo '<tr><td>Type</td><td>'.$details['type'].'</td></tr>'
-		.'<tr><td>Weight</td><td>'.$details['weight'].'</td></tr>'
-		.'<tr><td>Amount</td><td>'.$details['amount'].'</td></tr>'
-		.'<tr><td>Date</td><td>'.$details['date'].'</td></tr>';
+	$cusDetails = mysql_fetch_assoc(mysql_query("SELECT cus_id FROM customer_ref WHERE ref_no='$id'"));
+	$cusId = $cusDetails['cus_id'];
+	$cusDetails = mysql_fetch_assoc(mysql_query("SELECT name FROM customer_details WHERE cus_id='{$cusDetails['cus_id']}'"));
+	echo '<table width="120%" border="0">';
+	echo '<tr><td>Name</td><td>'.$cusDetails['name'].'</td></tr>'
+	.'<tr><td>NIC</td><td>'.$cusId.'</td></tr>'
+	.'<tr><td>Type</td><td>'.$details['type'].'</td></tr>'
+	.'<tr><td>Weight</td><td>'.$details['weight'].'</td></tr>'
+	.'<tr><td>Amount</td><td>'.$details['amount'].'</td></tr>'
+	.'<tr><td>Date</td><td>'.$details['date'].'</td></tr>';
 	echo '</table>';
 }
 ?>
